Rename volume_texture.h to texture.h
