+++ projects/PathosEngine/PathosEngine.vcxproj
@@ -184,7 +184,7 @@
     <ClCompile Include="src\pathos\material\material_shader.cpp" />
     <ClCompile Include="src\pathos\material\material_shader_assembler.cpp" />
     <ClCompile Include="src\pathos\rhi\shader_program.cpp" />
-    <ClCompile Include="src\pathos\rhi\volume_texture.cpp" />
+    <ClCompile Include="src\pathos\rhi\texture.cpp" />
     <ClCompile Include="src\pathos\text\font_mgr.cpp" />
     <ClCompile Include="src\pathos\text\font_texture_cache.cpp" />
     <ClCompile Include="src\pathos\text\text_geometry.cpp" />
@@ -324,7 +324,7 @@
     <ClInclude Include="src\pathos\rhi\shader_program.h" />
     <ClInclude Include="src\pathos\rhi\shader_storage_buffer.h" />
     <ClInclude Include="src\pathos\smart_pointer.h" />
-    <ClInclude Include="src\pathos\rhi\volume_texture.h" />
+    <ClInclude Include="src\pathos\rhi\texture.h" />
     <ClInclude Include="src\pathos\text\text_actor.h" />
     <ClInclude Include="src\pathos\text\text_component.h" />
     <ClInclude Include="src\pathos\util\engine_thread.h" />
