+++ projects/PathosEngine/PathosEngine.vcxproj.filters
@@ -264,7 +264,7 @@
     <ClCompile Include="src\pathos\render\shadow_omni.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="src\pathos\rhi\volume_texture.cpp">
+    <ClCompile Include="src\pathos\rhi\texture.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="src\pathos\render\depth_prepass.cpp">
@@ -707,7 +707,7 @@
     <ClInclude Include="src\pathos\render\shadow_omni.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="src\pathos\rhi\volume_texture.h">
+    <ClInclude Include="src\pathos\rhi\texture.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="src\pathos\render\depth_prepass.h">
