+++ projects/PathosEngine/src/pathos/loader/image_loader.cpp
@@ -1,6 +1,6 @@
 #include "pathos/loader/image_loader.h"
 
-#include "pathos/rhi/volume_texture.h"
+#include "pathos/rhi/texture.h"
 #include "pathos/util/resource_finder.h"
 #include "pathos/util/log.h"
