+++ projects/PathosEngine/src/pathos/render/volumetric_clouds.cpp
@@ -1,6 +1,6 @@
 #include "volumetric_clouds.h"
 #include "pathos/rhi/render_device.h"
-#include "pathos/rhi/volume_texture.h"
+#include "pathos/rhi/texture.h"
 #include "pathos/rhi/shader_program.h"
 #include "pathos/render/scene_proxy.h"
 #include "pathos/render/scene_render_targets.h"
